requirements.txt
fastapi
uvicorn
azure-storage-blob
azure-identity
PyPDF2
python-docx
openpyxl
httpx

